Refactored browser local storage into own function so that it can be used as hook or in background script
